FIRST COMMIT:
This is the first commit, and an essentially working version! Coded in a few hours, this is an actually working game that displays score for however long the player survives the onslaught. This will be my first game released to itch.io!

Eventually I'd like to add a main menu with a difficulty selector that changes speed and/or frequency of meteors. I'd also like to add sounds. But... FIRST GAME!!!