SECOND COMMIT: Additional features
This commit adds a difficulty feature, where the meteors get faster as time goes on. This also adds more player agency, where the player can on an interval shoot a bullet to destroy one meteor that is falling. This removes the meteor and the bullet from the screen.

However, there still is an issue where the bullets and meteors don't interact properly, and it seems to be from the array situation with removing them. I'll need to look into it and learn about removing array members, but the glitch is relatively minor, I think it shouldn't take too much to fix.
Once it is fixed, I can move this to itch.io and update the version of the game on there!