THIRD COMMIT: version 1.1.1 - RELEASED ON ITCH.IO
This commit fixes the bugs that plagued the 1.1.0 version where the meteors and bullet sprites would not properly disappear. The solution was, because of issues with remove from a nested for loop, they needed to be added to a separate list for removal which is cleared every frame. This seems to
have effectively solved the issue.

This version is ready for deployment and will be uploaded to itch.io. The .jar is updated and both are included, although the 1.0.0.jar got overwritten, so that will have to be re-archived.

Also updated the gitignore to not read anything from the archive folder.